requirements.txt
Flask==3.0.0
redis==5.0.1
sentence-transformers>=3.0.0,<4.0.0
PyPDF2==3.0.1
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
Werkzeug==3.0.1
